+++ packages/integration-sdk-runtime/src/synchronization/index.ts
@@ -9,6 +9,7 @@ import {
   Relationship,
   SynchronizationJob,
   IntegrationError,
+  IntegrationErrorEventName,
 } from '@jupiterone/integration-sdk-core';
 
 import { IntegrationLogger } from '../logger';
@@ -36,8 +37,14 @@ const UPLOAD_CONCURRENCY = 6;
 // 6291456 bytes, but we need header space.  Most web
 // servers will only allow 8KB or 16KB as a max header
 // size, so 6291456 - 16384 = 6275072
-// We're further reducing to a max size of 1 MB or 1048576
-const UPLOAD_SIZE_MAX = 1048576;
+// to be completely safe, we are using 6000000 bytes as default
+const MAX_BATCH_SIZE = 6000000;
+
+// TODO [INT-3707]: uncomment and use when implementing method
+// to shrink single entity's rawData until that entity is < 1MB
+
+// const MAX_RAW_DATA_SIZE = 1194304;
+
 export enum RequestHeaders {
   CorrelationId = 'JupiterOne-Correlation-Id',
 }
@@ -350,9 +357,8 @@ function handleUploadDataChunkError({
   );
 
   if (isRequestUploadTooLargeError(err)) {
-    logger.info(`Attempting to shrink rawData`);
-    const shrinkResults = shrinkRawData(batch);
-    logger.info(shrinkResults, 'Shrink raw data result');
+    // shrink rawData further to try and achieve a batch size of < 6MB
+    shrinkBatchRawData(batch, logger);
   } else if (systemErrorResponseData?.code === 'JOB_NOT_AWAITING_UPLOADS') {
     throw new IntegrationError({
       code: 'INTEGRATION_UPLOAD_AFTER_JOB_ENDED',
@@ -372,6 +378,8 @@ export async function uploadDataChunk<
 
   await retry(
     async (ctx) => {
+      // TODO [INT-3707]: on first try, shrink raw data of every entity
+      // to the point where each entity is < 1MB
       logger.info(
         {
           uploadCorrelationId,
@@ -443,14 +451,6 @@ interface KeyAndSize {
   size: number;
 }
 
-// Interface for shrink run results
-interface ShrinkRawDataResults {
-  initialSize: number;
-  totalSize: number;
-  itemsRemoved: number;
-  totalTime: number;
-}
-
 /**
  * Helper function to find the largest entry in an object and return its key
  * and approximate byte size.  We JSON.stringify as a method to try and have
@@ -475,42 +475,58 @@ function getLargestItemKeyAndByteSize(data: any): KeyAndSize {
 }
 
 /**
- * Helper function to find the largest Entity in our data array and return it.
+ * Helper function to find the entity in our data array with the largest rawData and return it.
  * We JSON.stringify as a method to try and have an apples to apples comparison
  * no matter what the data type of the value is.
  *
  * @param data
  * @returns
  */
-function getLargestEntityFromBatch(
+function getEntityFromBatchWithLargestRawData(
   data: UploadDataLookup[keyof UploadDataLookup][],
 ): Entity {
-  let largestItem;
-  let largestItemSize = 0;
+  let itemWithLargestRawData;
+  let largestRawDataSize = Number.MIN_SAFE_INTEGER;
+
+  for (const item of data) {
+    const length = item?._rawData
+      ? Buffer.byteLength(JSON.stringify(item._rawData))
+      : 0;
+    if (length > largestRawDataSize) {
+      itemWithLargestRawData = item;
+      largestRawDataSize = length;
+    }
+  }
+  return itemWithLargestRawData;
+}
+
+function getLargestEntityInBatch(
+  data: UploadDataLookup[keyof UploadDataLookup][],
+): Entity {
+  let largestEntity;
+  let largestEntitySize = Number.MIN_SAFE_INTEGER;
 
   for (const item of data) {
     const length = item ? Buffer.byteLength(JSON.stringify(item)) : 0;
-    if (length > largestItemSize) {
-      largestItem = item;
-      largestItemSize = length;
+    if (length > largestEntitySize) {
+      largestEntity = item;
+      largestEntitySize = length;
     }
   }
-  return largestItem;
+  return largestEntity;
 }
 
 /**
- * Helper function to find the largest _rawData entry in an Entity and return
+ * Helper function to find the largest element of the _rawData array in an Entity and return
  * it.  We JSON.stringify as a method to try and have an apples to apples comparison
  * no matter what the data type of the value is.
  *
  * @param data
  * @returns
  */
-function getLargestRawDataEntryFromEntity(
-  data: EntityRawData[],
-): EntityRawData {
+function getLargestRawDataEntry(data: EntityRawData[]): EntityRawData {
   let largestItem;
-  let largestItemSize = 0;
+  let largestItemSize = Number.MIN_SAFE_INTEGER;
 
   for (const item of data) {
     const length = item ? Buffer.byteLength(JSON.stringify(item)) : 0;
@@ -524,42 +540,93 @@ function getLargestRawDataEntryFromEntity(
 }
 
 /**
- * Removes data from the rawData of the largest entity until the overall size
- * of the data object is less than maxSize (defaulted to UPLOAD_SIZE_MAX).
+ * Helper function to generate a map of property sizes for a given Entity.
+ * This is used to determine what property keys on an Entity have a particularly large value.
+ * Intended to be used when logging upload errors due to large payload size.
  *
  * @param data
+ * @returns
  */
-export function shrinkRawData(
-  data: UploadDataLookup[keyof UploadDataLookup][],
-  maxSize = UPLOAD_SIZE_MAX,
-): ShrinkRawDataResults {
+function getPropSizeMapFromEntity(data: Entity): any {
+  const propSizeMap = {};
+
+  for (const [key, value] of Object.entries(data)) {
+    propSizeMap[key] = Buffer.byteLength(JSON.stringify(value));
+  }
+
+  return propSizeMap;
+}
+
+/**
+ * Removes data from the rawData of the largest entity until the overall size
+ * of the data object is less than maxSize (defaulted to MAX_BATCH_SIZE).
+ *
+ * @param batchData
+ */
+export function shrinkBatchRawData(
+  batchData: UploadDataLookup[keyof UploadDataLookup][],
+  logger: IntegrationLogger,
+  maxBatchSize = MAX_BATCH_SIZE,
+): void {
+  logger.info(`Attempting to shrink rawData`);
   const startTimeInMilliseconds = Date.now();
-  let totalSize = Buffer.byteLength(JSON.stringify(data));
-  const initialSize = totalSize;
+  let totalBatchSize = Buffer.byteLength(JSON.stringify(batchData));
+  const initialBatchSize = totalBatchSize;
   let itemsRemoved = 0;
   const sizeOfTruncated = Buffer.byteLength("'TRUNCATED'");
 
-  while (totalSize > maxSize) {
-    // Find largest Entity
-    const largestEntity = getLargestEntityFromBatch(data);
+  while (totalBatchSize > maxBatchSize) {
+    // Find Entity with largest rawData
+    const entityWithLargestRawData =
+      getEntityFromBatchWithLargestRawData(batchData);
 
     // If we don't have any entities to shrink or the optional _rawData array is empty,
     // we have no other options than to throw an error.
-    if (largestEntity?._rawData) {
-      // Find largest _rawData entry (typically 0, but check to be certain)
-      const largestRawDataEntry = getLargestRawDataEntryFromEntity(
-        largestEntity._rawData,
+    if (entityWithLargestRawData?._rawData) {
+      // Find largest element of the _rawData array (typically at index 0, but check to be certain)
+      const largestRawDataEntry = getLargestRawDataEntry(
+        entityWithLargestRawData._rawData,
       );
-
-      // Find largest item within rawData
+      // Find largest item within that element
       const largestItemLookup = getLargestItemKeyAndByteSize(
         largestRawDataEntry.rawData,
       );
 
+      // if we can no longer truncate, log and error out
+      if (largestItemLookup.size === sizeOfTruncated) {
+        const sizeDistribution = {};
+        for (const entity of batchData) {
+          sizeDistribution[entity._key] = Buffer.byteLength(
+            JSON.stringify(entity),
+          );
+        }
+        logger.error(
+          {
+            largestEntityPropSizeMap: getPropSizeMapFromEntity(
+              getLargestEntityInBatch(batchData),
+            ),
+            totalBatchSize: totalBatchSize,
+            sizeDistribution,
+          },
+          'Encountered upload size error after fully shrinking. This is likely due to properties on the entity being too large in size.',
+        );
+        logger.publishErrorEvent({
+          name: IntegrationErrorEventName.EntitySizeLimitEncountered,
+          description: `Failed to upload integration data because the payload is too large. This batch of ${batchData.length} entities is still ${totalBatchSize} bytes after truncating all non-mapped properties.`,
+        });
+        throw new IntegrationError({
+          code: 'INTEGRATION_UPLOAD_FAILED',
+          fatal: false,
+          message:
+            'Failed to upload integration data because the payload is still too large after performing as much shrinking as possible',
+        });
+      }
+
       // Truncate largest item and recalculate size to see if we need to continue truncating additional items
       largestRawDataEntry.rawData[largestItemLookup.key] = 'TRUNCATED';
       itemsRemoved += 1;
-      totalSize = totalSize - largestItemLookup.size + sizeOfTruncated;
+      totalBatchSize =
+        totalBatchSize - largestItemLookup.size + sizeOfTruncated;
     } else {
       // Cannot find any entities to shrink, so throw
       throw new IntegrationError({
@@ -572,12 +639,15 @@ export function shrinkRawData(
   }
 
   const endTimeInMilliseconds = Date.now();
-  return {
-    initialSize,
-    totalSize,
-    itemsRemoved,
-    totalTime: endTimeInMilliseconds - startTimeInMilliseconds,
-  };
+  logger.info(
+    {
+      initialSize: initialBatchSize,
+      totalSize: totalBatchSize,
+      itemsRemoved,
+      totalTime: endTimeInMilliseconds - startTimeInMilliseconds,
+    },
+    'Shrink raw data result',
+  );
 }
 
 interface AbortSynchronizationInput extends SynchronizationJobContext {