+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to
 
 ## Unreleased
 
+## [8.13.5] - 2022-05-11
+
+### Fixed
+
+- Hanging execution when encountering upload failure due to entity properties
+  being too large
+- add logging surrounding the size distribution of upload batch and its largest
+  entity when such a situation occurs
+- throw fatal error when such a situation occurs
+
 ## [8.13.4] - 2022-05-10
 
 ### Fixed